test(distill): consolidate 10 distill test files into 4 + shared fixtures
- Fold 7 small/topical files into 3 broader topical files plus a new
  config_cli file:
  - schema (+ artifact_schemas_distill_review)
  - gate (+ finalize_gate + closure_guards)
  - integration (+ candidates)
  - config_cli (config + cli + skill_bundle)
- Lift shared setup helpers (write_quest_yaml, make_quest_root,
  seed_runs, seed_pending_run, seed_worktree_pending_run,
  seed_distill_review, seed_distill_review_in_quest,
  make_artifact_quest, enable_distill_in_quest, make_decision_record,
  make_baseline_guidance) into tests/_distill_fixtures.py — eliminates
  ~200 lines of fixture duplication across 10 files
- All 102 test function names preserved; assertions byte-identical to
  pre-consolidation behavior; no test dropped
- Net: 10 files -> 4 + 1 helper; 2002 -> 1992 lines (line trim is small
  because savings from collapsed helpers were offset by the helper
  module + per-file docstrings; the visible-file-count drop is the
  win the PR reviewer was asking for)

Per PR feedback: 10 distill test files with duplicated fixtures was
test-file proliferation. Same coverage in fewer files.
